server/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
boto3==1.13.9
botocore==1.16.9
click==7.1.1
colorama==0.4.3
distlib==0.3.0
Django==3.0.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.17
Werkzeug==1.0.1
wrapt==1.11.2
